refactoring - move createFileInfo() and getItemInfosForUploaderFiles() methods form Editing to Controller
